CtrlLib: EditNumber fixed (thanks SenderGhost!) #491
git-svn-id: svn://ultimatepp.org/upp/trunk@6190 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/CtrlLib/CtrlUtil.cpp
@@ -186,7 +186,7 @@ bool EditNumber(int& n, const char *title, const char *label, int min, int max, 
 	return false;
 }
 
-bool EditNumber(int& n, const char *title, const char *label, double min, double max, bool notnull)
+bool EditNumber(double& n, const char *title, const char *label, double min, double max, bool notnull)
 {
 	WithEditDoubleLayout<TopWindow> dlg;
 	CtrlLayoutOKCancel(dlg, title);
